refactor: rename useAccordion folder to just hooks
